requirements.txt
click==8.0.1
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
PyMySQL==1.0.2
PyYAML==5.4.1
SQLAlchemy==1.4.21
Werkzeug==2.0.1
WTForms==2.3.3